+++ helper-scripts/autoupdate.sh
@@ -24,6 +24,7 @@ echo '#!/bin/bash' >> /tmp/ErsatzTV-Filler-autoupdate.sh
 echo cd $scriptdir >> /tmp/ErsatzTV-Filler-autoupdate.sh
 echo git pull >> /tmp/ErsatzTV-Filler-autoupdate.sh
 echo "rm -f $scriptdir/running.txt" >> /tmp/ErsatzTV-Filler-autoupdate.sh
+echo touch $workdir/update-run >> /tmp/ErsatzTV-Filler-autoupdate.sh
 echo exec $scriptdir/generator.sh >> /tmp/ErsatzTV-Filler-autoupdate.sh
 
 chmod +x /tmp/ErsatzTV-Filler-autoupdate.sh