+++ helper-scripts/channel-offline.sh
@@ -11,6 +11,11 @@ then
   output=/output
 fi
 
+if [[ -f $workdir/update-run ]];
+then
+processchanneloffline=no
+fi
+
 processchanneloffline1=$(echo $processchanneloffline | tr '[:upper:]' '[:lower:]')
 if [[ $processchanneloffline1 = yes ]]
 then